search: Move zoektSearchOpts to internal/search/zoekt
This moves a couple more things to internal/sesarch/zoekt to make them
available for searcher to use.
