+++ internal/search/zoekt/zoekt.go
@@ -1,10 +1,21 @@
 package zoekt
 
 import (
+  "context"
 	"regexp/syntax"
+  "time"
 
+  "github.com/google/zoekt"
 	zoektquery "github.com/google/zoekt/query"
+  "github.com/opentracing/opentracing-go"
+  "github.com/inconshreveable/log15"
+
+  "github.com/sourcegraph/sourcegraph/internal/trace/ot"
+  "github.com/sourcegraph/sourcegraph/internal/search"
 )
+
+const defaultMaxSearchResults = 30
+var defaultTimeout = 20 * time.Second
 
 func FileRe(pattern string, queryIsCaseSensitive bool) (zoektquery.Q, error) {
 	return ParseRe(pattern, true, false, queryIsCaseSensitive)
@@ -42,4 +53,47 @@ func ParseRe(pattern string, filenameOnly bool, contentOnly bool, queryIsCaseSen
 		Content:       contentOnly,
 		FileName:      filenameOnly,
 	}, nil
+}
+
+func getSpanContext(ctx context.Context) (shouldTrace bool, spanContext map[string]string) {
+	if !ot.ShouldTrace(ctx) {
+		return false, nil
+	}
+
+	spanContext = make(map[string]string)
+	if err := ot.GetTracer(ctx).Inject(opentracing.SpanFromContext(ctx).Context(), opentracing.TextMap, opentracing.TextMapCarrier(spanContext)); err != nil {
+		log15.Warn("Error injecting span context into map: %s", err)
+		return true, nil
+	}
+	return true, spanContext
+}
+
+func SearchOpts(ctx context.Context, k int, query *search.TextPatternInfo) zoekt.SearchOptions {
+	shouldTrace, spanContext := getSpanContext(ctx)
+	searchOpts := zoekt.SearchOptions{
+		Trace:                  shouldTrace,
+		SpanContext:            spanContext,
+		MaxWallTime:            defaultTimeout,
+		ShardMaxMatchCount:     100 * k,
+		TotalMaxMatchCount:     100 * k,
+		ShardMaxImportantMatch: 15 * k,
+		TotalMaxImportantMatch: 25 * k,
+		MaxDocDisplayCount:     2 * defaultMaxSearchResults,
+	}
+
+	// We want zoekt to return more than FileMatchLimit results since we use
+	// the extra results to populate reposLimitHit. Additionally the defaults
+	// are very low, so we always want to return at least 2000.
+	if query.FileMatchLimit > defaultMaxSearchResults {
+		searchOpts.MaxDocDisplayCount = 2 * int(query.FileMatchLimit)
+	}
+	if searchOpts.MaxDocDisplayCount < 2000 {
+		searchOpts.MaxDocDisplayCount = 2000
+	}
+
+	if userProbablyWantsToWaitLonger := query.FileMatchLimit > defaultMaxSearchResults; userProbablyWantsToWaitLonger {
+		searchOpts.MaxWallTime *= time.Duration(3 * float64(query.FileMatchLimit) / float64(defaultMaxSearchResults))
+	}
+
+	return searchOpts
 }
